test: consolidate pipeline runner fakes
